HTBHF-978 extract check post helpers to seperate files
