+++ src/web/routes/application/check/constants.js
@@ -1,0 +1,7 @@
+const CLAIMS_ENDPOINT = `/v1/claims`
+const NO_ELIGIBILITY_STATUS_MESSAGE = 'The claimant service did not return an eligibility status'
+
+module.exports = {
+  CLAIMS_ENDPOINT,
+  NO_ELIGIBILITY_STATUS_MESSAGE
+}
